don't block in poll longer than the autocommit/heartbeat intervals

+++ clients/src/main/java/org/apache/kafka/clients/consumer/ConsumerConfig.java
@@ -225,7 +225,7 @@ public class ConsumerConfig extends AbstractConfig {
                                         Importance.MEDIUM,
                                         ENABLE_AUTO_COMMIT_DOC)
                                 .define(AUTO_COMMIT_INTERVAL_MS_CONFIG,
-                                        Type.LONG,
+                                        Type.INT,
                                         5000,
                                         atLeast(0),
                                         Importance.LOW,
